+++ mlir/lib/Dialect/Rock/Transforms/AffixTuningParameters.cpp
@@ -1,8 +1,9 @@
 #include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/Rock/IR/GemmContext.h"
 #include "mlir/Dialect/Rock/IR/Rock.h"
+#include "mlir/Dialect/Rock/IR/RockGemmWrapperInterface.h"
 #include "mlir/Dialect/Rock/Passes.h"
 #include "mlir/Dialect/Rock/Tuning/ConvContext.h"
-#include "mlir/Dialect/Rock/Tuning/GemmContext.h"
 #include "mlir/Dialect/Rock/Tuning/GridwiseGemmParams.h"
 #include "mlir/Dialect/Rock/Tuning/UtilityParams.h"
 #include "mlir/Dialect/Rock/utility/loweringUtils.h"
@@ -53,62 +54,6 @@ private:
   void affixBackwardDataUtilityKernels(Conv2DBwdDataOp &op);
 };
 } // anonymous namespace
-
-static ConvolutionDims obtainConvDims(Operation *op) {
-  auto filterLayoutAttr = op->getAttrOfType<ArrayAttr>("filter_layout");
-  auto inputLayoutAttr = op->getAttrOfType<ArrayAttr>("input_layout");
-  auto outputLayoutAttr =
-      op->template getAttrOfType<ArrayAttr>("output_layout");
-
-  // Get shape of filter tensor.
-  auto filterType = op->getOperand(0).getType().template cast<MemRefType>();
-  ArrayRef<int64_t> filterShape = filterType.getShape();
-
-  // Get shape of input tensor.
-  auto inputType = op->getOperand(1).getType().template cast<MemRefType>();
-  ArrayRef<int64_t> inputShape = inputType.getShape();
-
-  // Get shape of output tensor.
-  auto outputType = op->getOperand(2).getType().template cast<MemRefType>();
-  ArrayRef<int64_t> outputShape = outputType.getShape();
-
-  int64_t y, x, ho, wo, hi, wi, k, c, n, g;
-  y = x = ho = wo = hi = wi = k = c = n = g = 0;
-
-  for (unsigned i = 0; i < filterLayoutAttr.size(); ++i) {
-    auto filterAttr = filterLayoutAttr.getValue()[i].cast<StringAttr>();
-    auto inputAttr = inputLayoutAttr.getValue()[i].cast<StringAttr>();
-    auto outputAttr = outputLayoutAttr.getValue()[i].cast<StringAttr>();
-
-    if (filterAttr.getValue() == "y") {
-      y = filterShape[i];
-    } else if (filterAttr.getValue() == "x") {
-      x = filterShape[i];
-    } else if (filterAttr.getValue() == "k") {
-      k = filterShape[i];
-    } else if (filterAttr.getValue() == "c") {
-      c = filterShape[i];
-    } else if (filterAttr.getValue() == "g") {
-      g = filterShape[i];
-    }
-
-    if (inputAttr.getValue() == "hi") {
-      hi = inputShape[i];
-    } else if (inputAttr.getValue() == "wi") {
-      wi = inputShape[i];
-    } else if (inputAttr.getValue() == "ni") {
-      n = inputShape[i];
-    }
-
-    if (outputAttr.getValue() == "ho") {
-      ho = outputShape[i];
-    } else if (outputAttr.getValue() == "wo") {
-      wo = outputShape[i];
-    }
-  }
-
-  return ConvolutionDims(y, x, ho, wo, hi, wi, k, c, n, g);
-}
 
 void AffixTuningParameters::runOnOperation() {
   func::FuncOp func = getOperation();
@@ -164,9 +109,7 @@ void AffixTuningParameters::affixBackwardWeightUtilityKernels(
   if (bitEnumContainsAll(features, GemmFeatures::mfma)) {
     OpBuilder b(op.getContext());
 
-    ConvolutionDims convDims = obtainConvDims(op);
-    GemmContext gemmSize =
-        GemmContext::fromConvolution(ConvOpType::BwdWeight, convDims);
+    GemmContext gemmSize = op.getGemmSize();
 
     auto gemmParams =
         op->getAttrOfType<XdlopsGemmParamsAttr>(op.getParamsAttrName());
@@ -193,8 +136,6 @@ void AffixTuningParameters::affixBackwardWeightUtilityKernels(
 template <typename T>
 void AffixTuningParameters::affixTuningParametersImpl(T &op) {
   OpBuilder b(op.getContext());
-
-  ConvolutionDims dims = obtainConvDims(op);
 
   std::string perfConfig;
   if (auto perfConfigAttr =
